+++ package.json
@@ -18,6 +18,7 @@
     "prettier": "^2.8.7",
     "react": "18.2.0",
     "react-dom": "18.2.0",
+    "sass": "^1.62.0",
     "typescript": "^5.0.4"
   },
   "devDependencies": {
