+++ src/components/layouts/Header.tsx
@@ -1,0 +1,17 @@
+import styles from '@/styles/header.module.scss';
+
+export default function Header() {
+  const handleLogout = () => {
+    // logout();
+    // resetUser();
+    // router('/login');
+  };
+  return (
+    <header className={styles.wrapper}>
+      <div className={styles.title}>Routing Practice</div>
+      <button onClick={handleLogout} className={styles.button}>
+        Logout
+      </button>
+    </header>
+  );
+}
