+++ src/components/layouts/SideBar.tsx
@@ -1,0 +1,12 @@
+import Link from 'next/link';
+import styles from '@/styles/sidebar.module.scss';
+
+export default function SideBar() {
+  return (
+    <nav className={styles.wrapper}>
+      <ul className={styles.ulWrapper}>
+        <li>hihi</li>
+      </ul>
+    </nav>
+  );
+}
