+++ src/styles/common.scss
@@ -1,0 +1,58 @@
+/* color */
+$smoke-white: #f5f5f5;
+$smoke-red: #ffacb7;
+$smoke-blue: #6886c5;
+$dark-blue: #4d6391;
+$smoke-yellow: #ffe0ac;
+$black: #000000;
+$white: #ffffff;
+
+/* size */
+$section-border-weight: 0.3rem;
+$header-height: 5rem;
+$body-height: calc(100vh - $header-height);
+$sidebar-width: 15rem;
+$body-width: calc(100vw - $sidebar-width);
+$radius: 20px;
+
+/* font */
+$default-font: 'Poor Story', sans-serif;
+$title-font-size: 2rem;
+$button-font-size: 1.5rem;
+$default-font-size: 1rem;
+$middle-font-size: 1.2rem;
+
+/* mixin */
+@mixin flex-center {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+}
+
+@mixin flex-space-between {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+}
+
+@mixin flex-column {
+  display: flex;
+  flex-direction: column;
+}
+
+@mixin clickable-text {
+  color: $smoke-red;
+  cursor: pointer;
+  &:hover {
+    color: $smoke-yellow;
+  }
+}
+
+@mixin clickable-box {
+  background-color: $smoke-red;
+  color: $smoke-blue;
+  cursor: pointer;
+  &:hover {
+    background-color: $smoke-yellow;
+  }
+}
